Neural_Perceptron_SV- Description
This program applies Perceptron learning to a set of data generated by the system to demonstrate Perceptron neural networks. The user is given the option to choose between using datasets from AND or OR, which can be correctly classified by Perceptron, or XOR, which cannot be classified by any Perceptron network in any form.